refactor: 🎨 move strategy class to sperate files
